fix: estilos de paginação corrigidos
