+++ services/auth/authPOST.ts
@@ -1,9 +1,9 @@
-import axios from 'axios'
+import { useApi } from '../../composables/axios'
 import { navigateTo } from 'nuxt/app'
 
 export const authPost = async (password, email) => {
     try{
-        const { data } = await axios.post("http://localhost:8080/api/v1/auth", {
+        const { data } = await useApi().post("/auth", {
             "email": email,
             "password": password
         }, {
@@ -15,4 +15,26 @@ export const authPost = async (password, email) => {
     } catch(error){
         navigateTo('/login')
     }
+}
+
+export const recoveryTokenAUTH = async (token) => {
+    const { data } = await useApi().post('/auth/validateRecoveryToken', {
+        "token": token
+    }, {
+        headers: {
+            'Content-Type': 'application/json'
+        }
+    })
+    return data
+}
+
+export const forgotPasswordAUTH = async (email) => {
+    const { data } = await useApi().post('/auth/forgotPassword', {
+        "email": email
+    }, {
+        headers: {
+            'Content-Type': 'application/json'
+        }
+    })
+    return data
 }
